requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.7
django-jquery==3.1.0
Django==4.1.3
Werkzeug==2.3.7
apscheduler==3.10.4
requests==2.31.0
bcrypt==4.0.1